[feat/aut722] moved Log into its own widget
